musicvid/requirements.txt
anthropic>=0.40.0
openai-whisper>=20231117
librosa>=0.10.0
moviepy>=1.0.3
click>=8.1.0
requests>=2.31.0
Pillow>=10.0.0
cairosvg>=2.7.0
numpy>=1.24.0
soundfile>=0.12.0
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0
tenacity>=8.2.0
pytest>=7.0.0
